requirements.txt
asgiref==3.4.1
black==21.10b0
click==8.0.3
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
pytz==2021.3
regex==2021.11.2
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
